(MOT-4299) refactor(e2e): isolate local dashboard layers
