+++ .github/benchmark-site/overview.js
@@ -124,25 +124,6 @@
     efficiencyStatusCaption: document.querySelector("#efficiency-status-caption"),
     lastUpdate: document.querySelector("#last-update"),
     localRunner: document.querySelector("#local-runner"),
-    localForm: document.querySelector("#local-run-form"),
-    localSubmit: document.querySelector("#local-run-submit"),
-    localCancel: document.querySelector("#local-run-cancel"),
-    localRunStatus: document.querySelector("#local-run-status"),
-    localRunError: document.querySelector("#local-run-error"),
-    localRunLogShell: document.querySelector("#local-run-log-shell"),
-    localRunLog: document.querySelector("#local-run-log"),
-    localCatalogIndicator: document.querySelector("#local-catalog-indicator"),
-    localCatalogStatus: document.querySelector("#local-catalog-status"),
-    localConnectionUrl: document.querySelector("#local-connection-url"),
-    localCatalogRefresh: document.querySelector("#local-catalog-refresh"),
-    localSubject: document.querySelector("#local-subject"),
-    localJudge: document.querySelector("#local-judge"),
-    localScenarioPicker: document.querySelector("#local-scenario-picker"),
-    localScenarioSummary: document.querySelector("#local-scenario-summary"),
-    localScenarioOptions: document.querySelector("#local-scenario-options"),
-    localScenarioAll: document.querySelector("#local-scenario-all"),
-    localScenarioNone: document.querySelector("#local-scenario-none"),
-    localAdvanced: document.querySelector("#local-advanced"),
     matrix: document.querySelector("#health-matrix"),
     next: document.querySelector("#next-page"),
     pageLabel: document.querySelector("#page-label"),
@@ -1319,300 +1300,11 @@
       state.page += 1;
       renderTable();
     });
-    if (isLocal) initializeLocalRunner();
+    if (isLocal) window.HarnessLocalRunner.initialize();
     render();
     await hydrateExecutionMetrics();
     renderEfficiency();
     renderTable();
-  }
-
-  let localPollTimer = null;
-  let localCatalogReady = false;
-  let localCatalogLoading = false;
-  let localJobActive = false;
-  let localDefaults = {};
-
-  function localFormField(name) {
-    return elements.localForm.elements.namedItem(name);
-  }
-
-  function applyLocalDefaults(defaults) {
-    localDefaults = { ...localDefaults, ...(defaults || {}) };
-    Object.entries(defaults || {}).forEach(([name, value]) => {
-      const field = localFormField(name);
-      if (field && !field.value && value !== null && value !== undefined) {
-        field.value = String(value);
-      }
-    });
-    const url = localFormField("url")?.value || defaults?.url || "";
-    if (url) elements.localConnectionUrl.textContent = url;
-  }
-
-  function setLocalControls(active) {
-    localJobActive = active;
-    for (const field of elements.localForm.elements) {
-      if (field !== elements.localCancel) field.disabled = active;
-    }
-    elements.localSubject.disabled = active || !localCatalogReady;
-    elements.localJudge.disabled = active || !localCatalogReady;
-    elements.localSubmit.disabled = active || !localCatalogReady;
-    elements.localCatalogRefresh.disabled = active || localCatalogLoading;
-    elements.localScenarioPicker.classList.toggle(
-      "local-picker-disabled",
-      active || !localCatalogReady,
-    );
-    elements.localScenarioPicker.setAttribute(
-      "aria-disabled",
-      String(active || !localCatalogReady),
-    );
-  }
-
-  function renderLocalJob(response) {
-    applyLocalDefaults(response?.defaults);
-    const job = response?.job;
-    const active = ["running", "cancelling"].includes(job?.status);
-    setLocalControls(active);
-    elements.localCancel.hidden = !active;
-    elements.localRunError.hidden = !job?.error;
-    elements.localRunError.textContent = job?.error || "";
-    elements.localRunLogShell.hidden = !job?.log;
-    elements.localRunLog.textContent = job?.log || "";
-    if (job?.log && active) elements.localRunLogShell.open = true;
-    elements.localRunStatus.textContent = !job
-      ? "Ready"
-      : {
-          running: "Running…",
-          cancelling: "Cancelling…",
-          cancelled: "Cancelled",
-          completed: "Results saved",
-          failed: "Runner failed",
-        }[job.status] || job.status;
-    if (active) {
-      clearTimeout(localPollTimer);
-      localPollTimer = setTimeout(refreshLocalJob, 1_000);
-    } else if (job?.status === "completed" && job.id) {
-      const reloadKey = "harness-e2e-local-last-reload";
-      if (sessionStorage.getItem(reloadKey) !== job.id) {
-        sessionStorage.setItem(reloadKey, job.id);
-        window.location.reload();
-      }
-    }
-  }
-
-  async function localApi(path, options = {}) {
-    const response = await fetch(path, {
-      cache: "no-store",
-      headers: { "Content-Type": "application/json" },
-      ...options,
-    });
-    const payload = await response.json().catch(() => ({}));
-    if (!response.ok) throw new Error(payload.error || `Request failed (${response.status})`);
-    return payload;
-  }
-
-  async function refreshLocalJob() {
-    try {
-      const response = await localApi("./api/local/run");
-      renderLocalJob(response);
-      return response;
-    } catch (error) {
-      elements.localRunError.hidden = false;
-      elements.localRunError.textContent = error.message;
-      elements.localRunStatus.textContent = "Unavailable";
-      return null;
-    }
-  }
-
-  function modelKey(model) {
-    return `${model.provider}\n${model.model}`;
-  }
-
-  function selectedModel(select) {
-    const option = select.selectedOptions[0];
-    return option?.dataset.model && option?.dataset.provider
-      ? { model: option.dataset.model, provider: option.dataset.provider }
-      : null;
-  }
-
-  function fillModelSelect(select, models, { includeAutomatic = false } = {}) {
-    const selected = selectedModel(select);
-    const preferredKey =
-      (selected && modelKey(selected)) ||
-      localStorage.getItem("harness-e2e-local-subject") ||
-      (localDefaults.model && localDefaults.provider
-        ? modelKey({ model: localDefaults.model, provider: localDefaults.provider })
-        : "");
-    select.replaceChildren();
-    if (includeAutomatic) {
-      const automatic = document.createElement("option");
-      automatic.value = "";
-      automatic.textContent = "Use subject model when required";
-      select.append(automatic);
-    }
-    models.forEach((model, index) => {
-      const option = document.createElement("option");
-      option.value = `model-${index}`;
-      option.dataset.model = model.model;
-      option.dataset.provider = model.provider;
-      option.textContent = `${model.provider} / ${model.model}`;
-      option.selected = !includeAutomatic && modelKey(model) === preferredKey;
-      select.append(option);
-    });
-    if (!includeAutomatic && select.selectedIndex < 0 && select.options.length) {
-      select.selectedIndex = 0;
-    }
-  }
-
-  function scenarioInputs() {
-    return [...elements.localScenarioOptions.querySelectorAll("input[type=checkbox]")];
-  }
-
-  function updateScenarioSummary() {
-    const inputs = scenarioInputs();
-    const selected = inputs.filter((input) => input.checked).length;
-    if (!inputs.length) {
-      elements.localScenarioSummary.textContent = localCatalogLoading
-        ? "Loading scenarios…"
-        : "Catalog unavailable";
-      elements.localSubmit.disabled = true;
-      return;
-    }
-    elements.localScenarioSummary.textContent =
-      selected === inputs.length
-        ? `All ${inputs.length} scenarios`
-        : `${selected} of ${inputs.length} scenarios`;
-    elements.localSubmit.disabled =
-      localJobActive || !localCatalogReady || selected === 0;
-  }
-
-  function fillScenarios(scenarios) {
-    const previous = new Set(
-      scenarioInputs().filter((input) => input.checked).map((input) => input.value),
-    );
-    const selectAll = previous.size === 0;
-    elements.localScenarioOptions.replaceChildren();
-    scenarios.forEach((scenarioId, index) => {
-      const label = document.createElement("label");
-      label.className = "local-scenario-option";
-      const input = document.createElement("input");
-      input.type = "checkbox";
-      input.name = "scenario";
-      input.value = scenarioId;
-      input.id = `local-scenario-${index}`;
-      input.checked = selectAll || previous.has(scenarioId);
-      const text = document.createElement("span");
-      text.textContent = scenarioId.replaceAll("_", " ");
-      text.title = scenarioId;
-      label.append(input, text);
-      elements.localScenarioOptions.append(label);
-    });
-    updateScenarioSummary();
-  }
-
-  async function refreshLocalCatalog() {
-    const url = localFormField("url")?.value || localDefaults.url || "";
-    elements.localConnectionUrl.textContent = url;
-    elements.localCatalogStatus.textContent = "Discovering the running Harness…";
-    elements.localCatalogIndicator.className = "local-connection-dot";
-    localCatalogLoading = true;
-    localCatalogReady = false;
-    setLocalControls(localJobActive);
-    try {
-      const query = new URLSearchParams({ url });
-      const catalog = await localApi(`./api/local/catalog?${query}`);
-      fillModelSelect(elements.localSubject, catalog.models);
-      fillModelSelect(elements.localJudge, catalog.models, { includeAutomatic: true });
-      fillScenarios(catalog.scenarios);
-      localCatalogReady = true;
-      elements.localCatalogIndicator.className = "local-connection-dot connected";
-      elements.localCatalogStatus.textContent =
-        `${catalog.models.length} registered model${catalog.models.length === 1 ? "" : "s"} · ${catalog.scenarios.length} scenarios`;
-      elements.localRunError.hidden = true;
-    } catch (error) {
-      elements.localCatalogIndicator.className = "local-connection-dot failed";
-      elements.localCatalogStatus.textContent = "Could not read the Harness catalog";
-      elements.localRunError.hidden = false;
-      elements.localRunError.textContent = error.message;
-      elements.localScenarioSummary.textContent = "Catalog unavailable";
-      elements.localAdvanced.open = true;
-    } finally {
-      localCatalogLoading = false;
-      setLocalControls(localJobActive);
-      updateScenarioSummary();
-    }
-  }
-
-  function initializeLocalRunner() {
-    elements.localForm.addEventListener("submit", async (event) => {
-      event.preventDefault();
-      const values = new FormData(elements.localForm);
-      const subject = selectedModel(elements.localSubject);
-      const judge = selectedModel(elements.localJudge);
-      const scenarios = scenarioInputs()
-        .filter((input) => input.checked)
-        .map((input) => input.value);
-      try {
-        if (!subject) throw new Error("Select a registered subject model.");
-        if (!scenarios.length) throw new Error("Select at least one scenario.");
-        localStorage.setItem("harness-e2e-local-subject", modelKey(subject));
-        elements.localRunError.hidden = true;
-        renderLocalJob(
-          await localApi("./api/local/run", {
-            method: "POST",
-            body: JSON.stringify({
-              label: values.get("label"),
-              url: values.get("url"),
-              model: subject.model,
-              provider: subject.provider,
-              judge_model: judge?.model || "",
-              judge_provider: judge?.provider || "",
-              scenarios,
-              runs: Number(values.get("runs")),
-              technical_retries: Number(values.get("technical_retries")),
-            }),
-          }),
-        );
-      } catch (error) {
-        elements.localRunError.hidden = false;
-        elements.localRunError.textContent = error.message;
-        elements.localRunStatus.textContent = "Could not start";
-      }
-    });
-    elements.localCancel.addEventListener("click", async () => {
-      try {
-        renderLocalJob(
-          await localApi("./api/local/run/cancel", {
-            method: "POST",
-            body: "{}",
-          }),
-        );
-      } catch (error) {
-        elements.localRunError.hidden = false;
-        elements.localRunError.textContent = error.message;
-      }
-    });
-    elements.localCatalogRefresh.addEventListener("click", () => {
-      refreshLocalCatalog();
-    });
-    elements.localScenarioAll.addEventListener("click", () => {
-      scenarioInputs().forEach((input) => {
-        input.checked = true;
-      });
-      updateScenarioSummary();
-    });
-    elements.localScenarioNone.addEventListener("click", () => {
-      scenarioInputs().forEach((input) => {
-        input.checked = false;
-      });
-      updateScenarioSummary();
-    });
-    elements.localScenarioOptions.addEventListener("change", updateScenarioSummary);
-    elements.localScenarioPicker.addEventListener("toggle", () => {
-      if (!localCatalogReady && elements.localScenarioPicker.open) {
-        elements.localScenarioPicker.open = false;
-      }
-    });
-    refreshLocalJob().then(() => refreshLocalCatalog());
   }
 
   initialize();
