+++ harness/tests/e2e/src/dashboard/assets.rs
@@ -1,0 +1,43 @@
+use axum::body::Body;
+use axum::http::{header, HeaderValue, StatusCode, Uri};
+use axum::response::{IntoResponse, Response};
+use rust_embed::RustEmbed;
+
+#[derive(RustEmbed)]
+#[folder = "../../../.github/benchmark-site/"]
+#[exclude = "*.test.cjs"]
+#[exclude = "README.md"]
+#[exclude = "sample-data.js"]
+#[exclude = "sample-executions.js"]
+struct DashboardAssets;
+
+pub(super) async fn static_asset(uri: Uri) -> Response {
+    let path = uri.path().trim_start_matches('/');
+    let path = if path.is_empty() { "index.html" } else { path };
+    let Some(asset) = DashboardAssets::get(path) else {
+        return StatusCode::NOT_FOUND.into_response();
+    };
+    let mime = mime_guess::from_path(path).first_or_octet_stream();
+    let mut response = Response::new(Body::from(asset.data));
+    response.headers_mut().insert(
+        header::CONTENT_TYPE,
+        HeaderValue::from_str(mime.as_ref())
+            .unwrap_or(HeaderValue::from_static("application/octet-stream")),
+    );
+    response
+        .headers_mut()
+        .insert(header::CACHE_CONTROL, HeaderValue::from_static("no-store"));
+    response
+}
+
+pub(super) fn javascript_response(body: String) -> Response {
+    let mut response = Response::new(Body::from(body));
+    response.headers_mut().insert(
+        header::CONTENT_TYPE,
+        HeaderValue::from_static("text/javascript; charset=utf-8"),
+    );
+    response
+        .headers_mut()
+        .insert(header::CACHE_CONTROL, HeaderValue::from_static("no-store"));
+    response
+}
